MotionGPT/requirements.txt
tensorboard
pytorch_lightning
torchmetrics
omegaconf
shortuuid
chumpy
diffusers
einops
wandb
rich
matplotlib
scipy==1.13.1

# for visualization
smplx==0.1.28
trimesh==3.9.24
joblib==1.2.0
h5py
scikit-image
spacy
ftfy
more-itertools
natsort
pyrender
moviepy
librosa
triangle==20220202

# additional dependencies for preference learning
trl==0.8.1
peft==0.10.0
colorlog==6.8.2
bert-score==0.3.13
opencv-python==4.9.0.80
shapely==2.0.3
numpy==1.23.1
bitsandbytes==0.43.0
transformers==4.38.2
accelerate==0.26.1
tyro==0.7.0